backend/requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
bcrypt==4.3.0
PyJWT==2.10.1
